refactor: change structure of game_library
