+++ game_library/src/spell_data.rs
@@ -11,6 +11,7 @@ use std::{any::Any, hash::Hash};
 use crate::{
     data_loader::DataFile,
     enums::{CastCategory, CastSlot, CastType, GameSystem, MagicType, Skill, SpellCollision},
+    shared_traits::KnownCastSlot,
     InternalId, StatEffect,
 };
 
@@ -104,6 +105,11 @@ pub struct SpellData {
     /// Debuffs that the spell can apply to the caster or to the target.
     #[serde(default = "Vec::new")]
     pub debuffs: Vec<StatEffect>,
+
+    // #### SPELL PARTICLES ####
+    /// Particles that the spell can create.
+    #[serde(default = "Vec::new")]
+    pub particles: Vec<SpellParticles>,
 }
 
 impl Hash for SpellData {
@@ -140,8 +146,33 @@ impl InternalId for SpellData {
     }
 }
 
-pub trait KnownCastSlot {
-    fn cast_slot(&self) -> CastSlot;
+/// A particle that a spell can create.
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SpellParticles {
+    /// The unique_id for the particle effect
+    pub particle_id: String,
+    /// The attachment point for the particle effect
+    pub attachment: ParticleAttachment,
+}
+
+/// The attachment point for a particle effect.
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub enum ParticleAttachment {
+    /// The particle is emitted once on the caster when the spell is cast.
+    OnCast,
+    /// The particle is emitted from the projectile while it is in flight.
+    OnProjectile,
+    /// The particle is emitted when the projectile hits something.
+    OnImpact,
+    /// The particle is emitted on the target when the spell hits.
+    OnTarget,
+    /// The particle is emitted on the caster when the spell is cast.
+    OnCaster,
+    /// The particle remains on the ground where the spell impacts.
+    OnGround,
+    /// The particle is emitted from the summoned entity.
+    OnSummon,
 }
 
 impl KnownCastSlot for SpellData {
@@ -255,6 +286,7 @@ impl Default for SpellData {
             angle: 0,
             buffs: Vec::new(),
             debuffs: Vec::new(),
+            particles: Vec::new(),
         }
     }
 }
